Fix heartbeat settings to work with ActiveMQ
ActiveMQ does not seem to respond to client-set heartbeats, but does
send heartbeats on a periodic schedule as requested at connect time.
This changes the heartbeat parameters such that the message broker's
heartbeat reat is greater than the client heartbeat rate, so that the
client always receives a heartbeat before it times out.